// web/src/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        JobDetail: {
            /** Format: uuid */
            readonly id: string;
            slug?: string | null;
            title: string;
            readonly company: components["schemas"]["CompanyNested"];
            location_raw?: string;
            work_mode?: components["schemas"]["WorkModeEnum"];
            job_type?: components["schemas"]["JobTypeEnum"];
            department?: string;
            /** Format: date-time */
            posted_at?: string | null;
            /** Format: uri */
            apply_url?: string;
            readonly is_saved: string;
            description?: string;
            readonly description_html_sanitized: string;
            /** Format: uri */
            source_url: string;
            experience_level?: components["schemas"]["ExperienceLevelEnum"];
            extraction_confidence?: number;
            /** Format: date-time */
            readonly scraped_at: string;
            readonly canonical_url: string;
            readonly structured_data: {
                [key: string]: unknown;
            } | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}